Reduces size of ORFNodeVector, removing TIS node storage and changing node coordinates to short unsigned ints.
Former-commit-id: a727f8a1aea4b1362040c6fb14246c4b62f759e1

+++ src/gene_graph.h
@@ -17,12 +17,12 @@ using namespace boost;
 using namespace std;
 
 // balrog scores for overlap penalties
-const double unidirectional_penalty_per_base = 3.895921717182765;  // 3' 5' overlap
-const double convergent_penalty_per_base = 4.603432608883688;  // 3' 3' overlap
-const double divergent_penalty_per_base = 3.3830814940689975;  // 5' 5' overlap
+const float unidirectional_penalty_per_base = 3.895921717182765;  // 3' 5' overlap
+const float convergent_penalty_per_base = 4.603432608883688;  // 3' 3' overlap
+const float divergent_penalty_per_base = 3.3830814940689975;  // 5' 5' overlap
 
 // typedefs for graph and edges for boost graph library
-typedef property<edge_weight_t, double> Weight;
+typedef property<edge_weight_t, float> Weight;
 typedef property<vertex_index_t, size_t> VertexIndex;
 typedef adjacency_list<vecS, vecS, bidirectionalS, no_property, Weight> GeneGraph;
 typedef std::pair<int, int> Edge;
@@ -50,10 +50,10 @@ protected:
 std::vector<size_t> getPath(
         const GeneGraph& graph,
         const std::vector<VertexDescriptor>& pMap,
-        const std::vector<double>& distances,
+        const std::vector<float>& distances,
         const VertexDescriptor& source,
         const VertexDescriptor& destination,
-        double& path_score,
+        float& path_score,
         const std::vector<size_t>& vertex_mapping);
 
 template <class T>
@@ -61,12 +61,12 @@ std::vector<size_t> traverse_components(const ORFVector& ORF_vector,
                                         const std::vector<size_t>& vertex_mapping,
                                         const std::unordered_set<size_t>& vertex_list,
                                         const GeneGraph& g,
-                                        const double& minimum_path_score,
+                                        const float& minimum_path_score,
                                         const size_t numVertices,
                                         T weight_pmap);
 
 std::vector<std::vector<size_t>> call_true_genes (const ORFVector& ORF_vector,
                                                   const ORFOverlapMap& overlap_map,
-                                                  const double& minimum_path_score);
+                                                  const float& minimum_path_score);
 
 #endif //GGCALLER_GENE_GRAPH_H
